[ENG-33] Apply `@tailwindcss/typography` to print layout
In [ENG-17], I couldn't make the Typography plugin work with Tailwind CSS v4. I presumed the plugin is not compatible with the new version of the library, and introduced my own `.prose` class as a temporary solution.

Now it turns out the plugin does work via the `@plugin` directive (instead of declaring it in `tailwind.config.ts`).

It seems that applying Tailwind's `prose` will require a bit of extra effort. Thus, in the scope of this task [ENG-33], I will use the plugin only for the printing layout.